docusense-functions/requirements.txt
azure-functions==1.15.0
azure-search-documents==11.*
azure-servicebus==7.11.0
requests==2.31.0
msal>=1.25.0
PyJWT==2.8.0
cryptography==41.0.8
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-pptx==0.6.23
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0 